refactor: :recycle: hook files for useCollapseProperties
